first draft of templates
