+++ components/Home.js
@@ -1,16 +1,20 @@
 import React from 'react';
 import { API, graphqlOperation } from 'aws-amplify'
 import { listCategoriess, getSubcategories, listSubcategoriess } from '../src/graphql/queries'
-import Album from './material-ui/Layout'
+import Landing from './material-ui/landing'
 
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
+// import CardMedia from '@material-ui/core/CardMedia';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
+
+import Header from './material-ui/headertemplate'
+import Footer from './material-ui/footertemplate'
+import AlgoTemp from './material-ui/algotemplate'
 
 class Home extends React.Component {
   constructor() {
@@ -49,7 +53,9 @@ class Home extends React.Component {
     console.log('cats', this.state.categories)
     return (
       <div >
-          <Album />
+        <AlgoTemp />
+        <Header />
+          <Landing />
           <br></br>
           {/* Categories Cards  */}
 
@@ -97,11 +103,7 @@ class Home extends React.Component {
 
   {/* Footer */}
   <br></br>
-  <footer>
-        <Typography variant="subtitle1" align="center" color="textSecondary" component="p">
-          PocketAlgo is the brainchild of Alex, Cesar and Vlad
-        </Typography>
-      </footer>
+          <Footer />
       {/* End footer */}
 
       </div>
