+++ components/material-ui/footertemplate.js
@@ -1,0 +1,55 @@
+import React from 'react';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+
+  icon: {
+    marginRight: theme.spacing(2),
+  },
+  heroContent: {
+    backgroundColor: theme.palette.background.paper,
+    padding: theme.spacing(8, 0, 6),
+  },
+  heroButtons: {
+    marginTop: theme.spacing(4),
+  },
+  cardGrid: {
+    paddingTop: theme.spacing(8),
+    paddingBottom: theme.spacing(8),
+  },
+  card: {
+    height: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  cardMedia: {
+    paddingTop: '56.25%', // 16:9
+  },
+  cardContent: {
+    flexGrow: 1,
+  },
+  footer: {
+    backgroundColor: theme.palette.background.paper,
+    padding: theme.spacing(6),
+  },
+}));
+
+export default function Footer () {
+  const classes = useStyles();
+
+  return (
+    <React.Fragment>
+  {/* Footer */}
+  <br></br>
+  <footer>
+        <Typography variant="subtitle1" align="center" color="textSecondary" component="p">
+          PocketAlgo is the brainchild of Alex, Cesar and Vlad
+        </Typography>
+      </footer>
+      {/* End footer */}
+
+
+    </React.Fragment>
+  );
+}
